subway-ops-monitor/requirements.txt
requests
supabase
python-dotenv
schedule
psycopg2-binary
pandas
